docker/app_layer/onchain-batch-txs/requirements.txt
pyarrow
pandas
fastparquet
pyspark
requests
delta-spark==3.1.0
pyiceberg[s3fs,hive]
hdfs
azure-identity
azure-keyvault
confluent-kafka
fastavro
duckdb